Make sure holes are printer after the roots
These constraints are actually too restrictive.

Let's say we have the following order constraints
```
(1_a > 2_a) (2_a > 3_a) (3_a > 4_a) (1_b > 2_b) (2_b > 3_b) (3_b > 4_b)
```
Here `a` and `b` are disjoint groups, and we would like them to be printed in a sequence. So print all `a` tool paths then all `b` tool paths or first all `b` then `a`. However as there is no inter-order constraints between `a` and `b`. The following is also a correct order
```
1a > 1b > 2a > 2b > 3a > 3b > 4a > 4b
```
This is solved by adding order constraints from the `a` order-tree to the root from `b`. Downside is that this removes the order optimization. If we have multiple holes we also add order constrains from one hole to the next hole. The ordering can later not be altered, so the order in which the holes are printed might not feel logical.

contributes to CURA-9551