'Workaround' for crash with tree-branch-pref set to 'flat surfaces'.
Added a note as well. I have no idea why the (in context) equivalent (even when the static_cast<size_t> is replaced by a simple cast to int) statement, using iota, doesn't work, but this does. It may just expose another bug, but best not to take the risk right now.

part of CURA-9879