Fix: Skirt was 'filling in' non-convex part of model.
The model (or rather, the 'dissalowed areas') gets subsstracted from the skirt/brim each offset. This caused an error since, for skirt, the offsets are calculated around the aproximate convex hull, but the part that was substracted from the area which should be included in each next pass was only the model itself. So, for a skirt with > 1 line-count, a skirt-distance low enough (approx. half a line width or less) to the model, and a model that was non-convex (concave), the skirt started filling in the non-conex 'gaps' left by the model and the convex skirt outline. This now no longer happens, since the covered and allowed areas are also 'aware' of the skirts 'convexity'.

CURA-9928