Add the initial visual_debugger class
Which is stored as a shared_pointer in the Application instance.

Still not satisfied with calling the writePVTu twice. But Since
I only know the number of generated files at the end, but the
folders etc, need to be created beforehand. I might create a
FR for this with them.

Because the Visual Logger has a custom destructor the rule of five
comes into play. Added all relevant, copy and move constructors
and operators.

Defined two basic concepts for isPolygon and isMesh, which certainly
needs to be refined since they are not generic enough, but this is mostly
due to how we setup our polygon class (not following standard conventions).

Added the vlogger ;-) to the ArcusCommunicationPrivate such that the mesh
will be written to the pvtu. Seemed like a logical place to start.

Contributes to CURA-10102