Optimize the log statements away in release mode
In debug mode it is enabled by default,
You would still have to set an environmental variable.
And in release mode you can add the preprocessor flag
`-DVISUAL_DEBUG` to still enable it.

The conan install command needs to be called with
the debug build_type or with the `enable_visual_debug=True`
to enable the vtu11 dependency and enable the option
for cmake.

Note: If you build en the debug and release on the same
computer, the same conan_toolchain.cmake will be used.
I need to figure out away to specify this per build_type
and if that can't be done, open an issue on the conan repo.

Contributes to CURA-10102