avoid support folding around model.
To avoid that the support is folding around the model, the support horizontal expansion should not cause the support to grow towards the model. Stepwise applying the support horizontal expansion to both the model outline and the support is effectively calculating a voronoi. The offset is first applied to the support and next to the model to ensure that the expanded support area is connected to the original support area. Please note that the horizontal expansion is rounded down to an integer offset_per_step.

CURA-10152