Move 'hole horizontal expansion' code.
Otherwise, fiddling with the holes happened _after_ the generation for interlocking structures, which is a process that can generate 'extra' holes in the structure meant for interlocking, that _shouldn't_ be affected by this parameter. Also, it's now in the same spot as where the other xy-offsets are done, which makes for less scattershot code.

should fix jira ticket CURA-10173