Move prime-tower-brim code when otherwise using skirt adhesion.
Brim and skirt use a mutually exclusive 'covered areas'. For the skirt, it's important to do a convex hull (-ish) operation on the initial covered areas. This shouldn't be done for brim. This caused trouble in the case that brim and skirt where used together, namely, when the prime tower brim is enabled, but the adhesion is otherwise skirt. This caused the brim of the prime-tower to adhere to the convex hull like covered areas for the skirt of the rest of the models, causing a gap in the prime-tower brim. The solution is mostly just being very careful when to add what, made easy by the fact that the machine and dissalowed-per-extruder areas don't need to be recalculated when (just) a convex hull operation is done afterwards.

should fix jira ticket CURA-10148