Use `min_wall_line_width` for sanitizing
Some models still wouldn't slice, by increasing the open-close
distance some more and relating it to the `min_wall_line_width`
(which makes more sense imo in this context) We finally got
in a state that all reported crashes should be fixed.

The simplification that this open-close operation performs should
well be within the limits of the printable features. No degradation
of quality should occur.

Fixes CURA-9976