Skin exposed to air should not be affected by interlocking.
Previously, 'thin strips' of material where expanded everywhere, to counteract the air-filtering from interfering with the adhesion to thin strips to the microstructures. This, however, also affected parts of the model where interlocking was not done, like skin. For these cases, yet another morphological operation was added to the filtering of thin parts, so now thin strips of material are only expanded into the other material _if there is actual interlocking microstructure cells to adhere to_.

This _still_ doesn't _completely_solve the problem, but should be good enough for the beta (another ticket has been made to resolve this completely).

part of CURA-10050