Prevent interlock workaround from intersection with complete model border.
Previously, as a workaround for interlocking structures not adhering to thin strips of material, an extra blob of material was added wherever interlocking microstructure was disconnected from the border-material. It turns out that this blob can intersect the border, completely negating (the purpose of) the air-boundary (the air-boundary is why these unexpected gaps can be there in the first place).

Now, we detect which parts of the border we can actually blob the connecting material on to. Because it does morphological operations in a loop, it might be slow. This will however be limited to whenever interlocking is on, and this workaround is expected to become a real fix before the final is out.

part of CURA-10050