Fix bug in generating support towers
Relax conditions used to add support towers as previously the condition for adding towers
was too strict. The routine works as follows: for each support polygon part that gets too
small (due to xy-distance from the model) a tower is created if the polygon part adheres
to the two conditions.
  - there is no model below the support area, and
  - there is support above the support area.
However, rather than doing a polygon-intersection check the center of the axis aligned
rectangle, bounding the small support area is taken for this check. If the center is both
inside the area of the support above, and _not_ in the area of the model below then a
tower is added. Taking the center of the aabb should not be an issue as the support area
is already small. However, there proved to be a difference between these two checks.
Solution was to do a proper intersection test rather than the center of aabb intersection
test.

There should _not_ be a significant hit on performance as this check is only executed once
for the tip of each candidate tower.

Contributes to CURA-10220