Fix bug in use towers
The `has_model_below` was checking in the horizontal disallowed areas around the model. This check proved to be too strict; when the center of the support tower is within the disallowed areas we might still want to place a support tower there as the support tower will grow beyond this disallowed area.

Contributes to CURA-10220