Move removal of xy_disallowed_areas to the main loop
This prevents jittering as layers sudden changes in disallowed can only change the support once. Previously the jittering in the disallowed areas would only effect the current layer, and not propagate downward causing the same jittery shapes to appear in the final support structure.

Additionally a procedure is added to the end of the support generation that removes floating support. Here all support is removed that has nothing above or below it.

Pay special attention to the comment
    // This is done after the main loop, because at least one of the calculations there rely on other layers _without_ the x/y-disallowed area.
and then choose to ignore it instead :gun:, (nah kidding this is a ductape solution which should be reverted once the jittery support is properly solved).

Contributes to CURA-10220