Remove spurious 'compensation'.
The distance between the raft and the first layer should remain the same, regardless of the 'layer 0 overlap', for the settings as currently written. Wrongly compensating it like this leads to it always reporting 1 layer or a gap, even though both the air-gap and the 'layer 0 overlap' are both 0. Just remove it then.

part of CURA-9860