Fix support generation for models with narrow walls
This was basically a side effect from CURA-10314. For this feature we were looking at the layer(s) above. The idea is that the overhang needs to exist for multiple layers before support is generated. When looking above the overhang angle is limited by the area of the above layer shrunken by `tan(overhang_angle) * delta_z`. As this shrinking behavior is achieved using a negative offset narrow walls would completely be removed.

In order to keep in line with the spirit of the contributions from CURA-10314 we instead look below rather then above. When looking below a positive offset need to be taken in order to achieve a similar result. As such features can never be removed.

CURA-10336