Speed up skin-edge support detection for detailed models.
Limit the considered areas to those that actually have infill to make 'infill skin edge support' from on the layer currently under consideration for the area determination. This can speed up the calculation for certain models a huge amount (under a minute of slicing versus half an hour kind of ballpark). Important, since we enabled skin edge support in the front-end by default instead of as an option.

CURA-9463