Resolve bug where inner polygon parts were removed if horizontal expansion is set to > 0
The hole removal was incorrectly applied on the resulting polygon in the iteration. When polygon `a` was encaptulating polygon `b` _and_ polygon `b` is first in the part-list then the polygon `b` would be removed as it hole of `a` would completely remove `b` by the difference operation.

Fixes #14764

CURA-10367