Add CLI-options to make 'lenient in what we accept' optional.
In the previous commit, the parsing of settings was made a bit more lenient. After discussion, we all agreed that fail fast is probably better: 1) We should not defing settings that are parents (non-leaf), but also hold values that should directly be read by the enigne. 2) There should always be a default value to parse, becasue we can never guarantee that the value is not a (python) formula. -- However, it's probably better if there's still the _option_ to do it this way, should something be messed up.

When developing this, some bugs where found (a bit too close to properly split it up in different commits) w.r.t. handling (search) directories. (Especially on Windows.)

done as part of CURA-9495