Unioning many (often small) circles took too long. Fix that.
The cipper union algorithm seems to be written for a small number of complex polygons, rather than a large number of simple ones. This workaround seems to work very well in case there is, say, a large number of mostly overlapping circles that need to be unioned that where added to a single Polygons object.

CURA-10464