Allow for default behaviour if no plugin is used
This commit updates the plugin registration in the Application class in the Application.cpp file. The changes include:

Adding support for default functions in the simplify_slot and postprocess_slot classes.
Creating a slot_registry typedef to hold the slots for simplification and post-processing.
Conditionally registering the simplify_slot based on whether the front-end starts a plugin.
Registering the postprocess_slot unconditionally.
Demonstrating the usage of the simplify_slot by invoking it with sample data.
These changes allow for flexible plugin registration and usage in the application.

Note: Further modifications may be required based on the specific requirements and configuration of the application.

[CURA-10475]