Prerequisites for secure validator.
Have secure communication when enterprise compile flag is set. The engine and plugins will both be started with the same shibolet, a plugin will then hash this with an openly avialable public key, so the engine can check that with the private key. It's assumed that the key-pair will only have to change when there has been a security violation.

part of CURA-10475