Update protobuf and grpc files
This commit updates the protobuf and grpc files to reflect the changes made on the arcus_replacement branch of the CuraEngine_grpc_definitions repository. The changes were necessary to ensure compatibility and consistency with the updated definitions.

This branch now uses the by @casperlamboo proposed new Protobuf types, but modified, such
that we can start pulling out libArcus and pyArcus.

Codebase consistency: The modifications made in this commit ensure that the protobuf and grpc files used in the current branch align with the definitions present in the arcus_replacement branch of the CuraEngine_grpc_definitions repository.

[CURA-10475]