Simplify: Iterate until it doesn't change anymore.
Simplify should be idempotent if the polygon already meets the criterea. Simplifying again shouldn't yield new results. The old version could do that, both in the sense that the by_importance vector wasn't completely exhausted, which could happen if if the poly[gon|line] was still larger than [3|2], since not every remove call removed a polygon. But also in the sense that iteration made it remove more in some cases. As these cases are not _that_ common, will only simplify the already simplified polygon, and simplification in and of itself can give speed boosts in the other part of the code (as well as the fact that this can't loop forever, sicne at some point there has to be an iteration where nothing is removed anymore), I think this is acceptable, at least for now.

part of CURA-10500