Smooth operator
Add a smooth action, which will smooth small segments (maximum resolution) of a polygon,
whom deviate more than a certain angle (wall transition angle) compared to a fluid motion
line between the segment before and after.

Smoothing is done by moving the first point of the small middle segment to the first point
of the previous point with a smooth distance (need to add this to the front end).
The second point of the middle segment is moved to the second point of the last segment.
If either the previous or next segment is shorter then the smooth distance, then that point
in the middle segment is removed

Contributes to CURA-10724
Contributes to Ultimaker/Cura#14811