Optimize smoothing function and fix error in shift_points
The smoothing function has been optimized and a bug with the
shift_points function has been fixed. The operator in the smoothing
function has been made "constexpr" to potentially improve performance.
The way "windows" are generated has been restructured to make it clearer
and eliminate redundancy. In shift_points, the subtraction used to alter
point's coordinates was replaced with addition to rectify a mistake. The
distance calculation in the smoothing function has been adjusted from an
incorrect dot product into a correct hypotenuse calculation. These
changes should help in improving the efficiency and correctness of the
smoothing function.

Contributes to CURA-10724