Add functionality to access 2D and 3D point coordinates by index or character
This commit addresses the need of accessing point coordinates more
flexibly. It introduces functions to access the coordinates of 2D and 3D
points either by coordinate index (e.g. 0 for X, 1 for Y in a 2D point)
or by character (e.g. 'x' or 'X' for X-coordinate). It makes handling
points more intuitive and comes in handy when the specific type of point
(2D or 3D) is not of interest. Additionally, accessing coordinates of a
junction by index or character is also implemented, further enhancing
the ease of use.

Contributes to CURA-10724