Update comment syntax in geometry.h
In this commit, the syntax for the comments in the geometry.h file has
been updated. The "\\" syntax was replaced by "@" as the latter is more
standard for Doxygen usage. This change allows for better parsing and
generation of documentation by Doxygen, and aligns with common practices
in the software industry.

Contributes to CURA-10724