Refactor smoothing function, remove unused parameter and improve readability
The smoothing function in `smooth.h` was refactored to improve code
readability and maintainability. The 'smooth_distance' parameter was
removed as it was not being utilized. The core logic of the function
remains unchanged, and several variable names were updated to better
reflect their context and function. Introduced a clearer, more accurate
method to calculate the cosine of the angle deviation which is also
generally optimized for smaller values of the `fluid_angle`. This is a
more computational efficient approach since for 'Curved' interpolations,
the `fluid_angle` is usually small which solves the primary problem.

Contributes to CURA-10724