Refactor smoother functionality in WallToolPaths
Moved the 'smoother' logic from the 'slicer' and 'WallToolPaths'
execution to occur earlier in the 'prepared_outline' execution. This
change intentional to reduce redundancy and improve the efficiency of
our slicing process by ensuring we're smoothing the polygons just once.
This could also help reduce the amount of crashes in the Voronoi
generation/usage

An unnecessary smoother function call with redundant parameters was also
removed from src/slicer.cpp."

Contributes to CURA-10724