Update smoothing parameters in test
Updated the parameters in the SmoothTest of CURA utility tests.
The third parameter of the smoothing function has been removed due to
recent changes in the CURA engine algorithm. Consequently, the expected
results of test were also updated to match the output from the updated
smoothing function.

Contributes to CURA-10724