Adjust Arachne type checks for different C++ versions
This commit addresses potential compatibility issues with different
versions of C++ by adding conditionals to the `junction` concept in the
`arachne.h` utility file. As some versions may not support the
`std::integral` check, we fall back to directly assessing `val.w` if
the C++ version is older or a particular library version with
compatibility issues is detected. This change ensures successful
compilation across varying environments.

Contributes to CURA-10724