Refactor std::integral usage for C++ compatibility
This commit removes the preprocessor conditions that were originally put
in to cope with an issue in Clang13 C++20. A fallback version of the
integral concept has been added in the std namespace if the C++ version
is below 201703L or if the library version is lesser than the Clang13
version. This allows for a cleaner and more compatible codebase across
various C++ versions. In addition, "include/utils/types/generic.h" is
now being included in "include/utils/types/geometry.h",
"include/utils/types/arachne.h" and "include/utils/actions/smooth.h"
everywhere std::integral might be used.

Contributes to CURA-10724