Less aggressive smoothing
The filter condition has been altered to be less aggressive, since the
intend is to smooth out the small sudden deviations. With this in mind
the removal of points is also discarded since this can introduce self-
intersecting shapes, which will result in hard crashes, but more
importantly introduce to much of a deviation from the intended shape.

Contribute to CURA-10724