Improve code readability and add assertions
Applied review comments.

This commit addresses several changes in the code to improve readability
and add more assertions where necessary. Redundant comments have been
removed, such as the namespace declaration in geometry.h for
readability. Grammar mistakes in comments were corrected for the same
reason in smooth.h.

In get.h and generic.h, a static assertion was added to restrict the
size of the string value to one character, ensuring concise point
coordinate inputs. In generic.h, a comment was added to indicate why we
are adding a manual implementation of integral and floating point, due
to lack of support in older versions of Apple Clang.

These changes were made to improve the quality and maintainability of
the code.

Contributes to CURA-10724