Move generic.h to types folder and add grpc_convertable concept
The file generic.h was moved from the utils/concepts directory to the
utils/types directory to better reflect the contents of the file, now
including a new concept for grpc_convertable. This new concept checks if
a type T is a semiregular type that can convert to grpc. This enhances
the code's clarity and can improve the development of grpc-utilizing
functionalities.

Contributes to CURA-10475