Shift UUID generation from FffGcodeWriter to Application
The changes are intended to centralize UUID generation inside the
Application class rather than in each instance of the FffGcodeWriter
class. This approach allows plugins to hook up with multiple instances
and still differentiate where the data comes from.
This is done by introducing instance_uuid data member in the Application
class which generates and stores UUID. This UUID is then used in
 FffGcodeWriter constructor. This change also removes the dependency on
  boost/uuid in FffGcodeWriter.

  Contributes to CURA-10715