Refactored co_spawn lambda in pluginproxy.h
Refactored the lambda function present in co_spawn calls located in
methods of pluginproxy.h. This change intends to increase readability
and maintenance. The lambda implementation was moved into private member
functions 'handshakeCall' and 'modifyCall'. This enhances code
organization allowing for easy modification and reuse in the future.

Contribute to CURA-10714