Moved private members in pluginproxy.h
This refactoring aims to make the public API more clear and
understandable by moving private members to the bottom of the file. With
this change, it will be easier for the developers to identify and
understand the public methods and attributes of the class at a first
glance, improving the maintainability and readability of the code.

Contributes to CURA-10714