Renamed operator() method to modify in plugin proxies
The operator() method was unclear in its intentions and use when applied
to processing plugin proxies. It's been renamed to 'modify', making it
clear that it's purpose is to alter or modify the input parameters
passed to the plugin. This affects all references to these proxies
within the code, improving overall readability.

Contribute to CURA-10805