Remove the previous way of handling slots without modify.
One one hand, this was arguably a hack, since it sort-of depended on the compiler throwing away the call to (plugin-proxy) modify, which, if not done, would eventually have lead to a compiler-error (in the situation before the plugin-proxy had a separated-out modify component at least), since NoStub doesn't have the requisite types associated with it to go through with the entire modify call-stack. On the other, the whole modify-component (in the PluginProxy file) despite not being a hack, has its own problems at the moment (overreliance on 'parent' class breaking modularization) and while perhaps more 'correct' does feel like a more cumbersome solution in way.

On the up side, this does give a bit more flexibility for coming up with different 'defaults' under different conditions. For when the behaviour should be slightly different when there is a plugin that doesn't implement one of the methods that is called versus no plugin being subscribed to that slot.

part of CURA-10851