Move slot generation step in infill switch
This was needed because otherwise, if a infill generate slot was loaded _all_ infill would automatically be routed through the plugin, bypassing the regular engine behavior. By executing the slot within the infill pattern type switch the plugin infill pattern generation is only executed when that infill pattern is set to plugin.

CURA-10619