Send allong flattend mesh settings
In this commit, the functionality to transfer flattened mesh settings
to plugins was added. This was necessary to ensure additional mesh
settings could be understood and
implemented by the plugins in the system.

A new utility function "getFlattendSettings" that returns flattened mesh
settings was introduced in the settings module. This functions gathers
settings from the parent containers as well as child containers for
completeness.

Finally, changes were made to the infill operation to adapt it to the
new functionality. The flattened mesh settings are now sent to plugin
function calls.

Contributes to CURA-10619