Made prep_client_context function static in common.h
This change was to ensure that the prep_client_context function isn't
exposed to the linker. This is a part of code clean-up as well as to
avoid any potential issues of function collision during linking. It
keeps this function limited to its translation unit, thereby enhancing
modularity and reducing possible side-effects. CURA-10851 relates to the
 task of improving code quality and maintainability.

 Contributes to CURA-10851