Refactor numeric types for better type safety and reduced code repetition
Numeric data types in the codebase are now defined as a numeric facade, providing greater type safety and reducing code repetition. This encompasses the `Acceleration`, `Velocity`, `Ratio`, and `LayerIndex` types.

This change updates the constructor and assignment operators, and ensures the facade can be used in a consistent manner with standard arithmetic types. It also supports user-defined type conversions to provide interoperability with built-in types.

In addition, correct type suffixes are added to numeric literals in a number of places to avoid potential bugs due to unexpected promotions or type coercions.

As part of this refactor, we also standardized the category of numeric types applied throughout the codebase, ensuring compatibility with these new types.

This refactor is needed to prep the GCodePaths/GCodePathConfigs for the plugin system. Since the changes for this refactor all over the placed, decided to make a seperate branch against main.

Contributes to CURA-10466