Update fluid motion mesh fix logic
This includes updated fluid motion logic. A new definition of fluid motion is introduced, here a tool-path-part is considered to be fluid if all of the following criterial hold
a) Vectors [A,B] and [C,D] are "long" (settable through settings)
b) Vector [B,C] is "short" (settable through settings)
c) either on the angles between vectors
    - [A_,D_] and BC or,
    - [A_,B] and BC or,
    - [B,C] and [C,D_]
    is below a threshold (settable through the settings)

                 B--C
                /    \
               A_     D_
              /        \
             /          \
            A            \
                          \
                           D

Unit tests are added, both to show intent of the `isSmooth` predicate and to test against future regressions.

CURA-10811