Update gcodeanalyzer.yml to fetch specific branch
Updated the 'gcodeanalyzer.yml' workflow to fetch files from the same branch in Cura repository if it exists. Previously, the script always pulled from the 'main' branch. The script now checks if the branch exists in the remote repository as well before pulling the '.def.json' files. This change allows for better synchronization between the branches of different repositories.