Update converter and proxy classes to include name and version of plugins
Modified the operator() in converter classes to include arguments for plugin name and version. Updated SlotProxy and PluginProxy constructors to accommodate these changes.
Also updated the handshake request to include the plugin name and version in its request. This change is necessary to facilitate identification and version control for plugins during the handshake process.

Contributes to CURA-10619