Also possible to exclude bottom from small skin behaviour.
Small skin areas replace the normal fill pattern and use walls. Previously in this branch, the possibility was added to not do that to areas exposed to air. It turns out this means in general, not just on the buildplate (also, bottoms can happen away from the buildplate as well of course). As such, take skin bottoms into account of when to exclude areas for consideration when 'Small Top/Bottom On Surface' is off.

part of CURA-10829