Refactor `MeshPathConfigs` related code for better readability
This commit refactors the `MeshPathConfigs` related code for clarity and
readability. It eliminates the `PathConfigStorage::` scope resolution
from `MeshPathConfigs` usages and refactors the SpeedDerivatives struct
initialization to use designated initialization. It also refactors the
`handleInitialLayerSpeedup` function to use method `smoothSpeed` on
`speed_derivatives` directly. This makes the code cleaner and easier to
understand.

Contribute to CURA-10446