Add conditional compilation for coroutines support
This commit updates the file `pluginproxy.h` to include support for both standard and experimental versions of coroutine. The preprocessor directives check if the standard coroutine header is available, if not it checks for the availability of the experimental version. In the latter's case, a flag `USE_EXPERIMENTAL_COROUTINE` is set. This change ensures compatibility across different compiler versions.

Contribute to CURA-10475