Improve and refactor the code for better readability
The commit drastically changes the way things are written in the code in
different files to improve readability and efficiency. Major changes
include an updated method for assigning properties to GCodePath and
MeshPathConfigs objects, using struct-designated initializers to
improve readability. Variable assignments were also refactored to make
understanding the code easier. Overall, the commit aims to make the code
more efficient and maintainable.

Contributes to CURA-10446