Large wasn't large enough.
The smaller value caused rounding errors still. Since coord_t should always be 64-bit, even if people compile it for 32-bit systems (which we don't officially support, but still...), it's safe to go to the 32-bit max. This will solve most or all of the remaining issues with the infill-line-crossing-resolve. (It's a bit more important than the previous commit in that regard.)

part of CURA-10410 second edition