Speed-up by using raw pointers (unfortunately).
Not a great solution, but it works. -- Creating a shared pointer via make_shared will copy the entire thing. Usually not so bad, but when it basicallly represents _an entire layer_ which is stores its member-variables by value, it can get ugly fast.

one proposed solution to internal ticket CURA-11019